terraform: ensure attached states use exact resource address match
The current AttachStateTransformer does a loose search for the
resource to attach using StateFilter.Filter() and the resource
address, taking the first match found. This can cause problems where
the state being searched for is a root module resource address and
there are other resources with the same name in other modules in the
state. In this case, the first match may not be the correct state.

This bug is pretty obscure and I would imagine hard to observe as the
only real repercussions I have personally seen is a removal of state
dependencies (ie: Dependencies, or depends_on in the state itself,
not config) during a refresh or plan. Specifically, this came up in
Sentinel testing, while checking the depends_on field in the tfstate
import for 0.11 states.

This commit fixes things by checking the results for an exact match.